+++ app/Services/User/Settings/SettingService.php
@@ -7,7 +7,7 @@ use App\Helpers\VerifyEmail;
 use App\Mail\User\SuccessUpdateEmailMail;
 use App\Mail\User\VerifyNewEmailUserMail;
 use App\Models\Users\ChangeEmail;
-use App\Models\Users\SpecificData;
+use App\Repositories\User\Settings\SettingRepository;
 use App\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
@@ -16,14 +16,41 @@ use Illuminate\Support\Str;
 class SettingService
 {
 
-    public function updateBasicData(array $data, SpecificData $item)
+    /**
+     * @var SettingRepository
+     */
+    private $settingRepository;
+
+    /**
+     * SettingService constructor.
+     * @param SettingRepository $settingRepository
+     */
+    public function __construct(SettingRepository $settingRepository)
     {
+        $this->settingRepository = $settingRepository;
+    }
+
+    /**
+     * @param array $data
+     * @return mixed
+     * @throws \Exception
+     */
+    public function updateBasicData(array $data)
+    {
+        $item = $this->settingRepository->findSpecificDataUser();
+        if (empty($item))
+            throw new \Exception(sprintf('Try again'));
         $item->update($data);
         return $item;
     }
 
-    public function tryUpdateEmailUser(array $data, $item)
+    /**
+     * @param array $data
+     * @return mixed
+     */
+    public function tryUpdateEmailUser(array $data)
     {
+        $item = $this->settingRepository->findTmpChangeEmail();
         $data['_key'] = VerifyEmail::generateKey();
         if (empty($item)) {
             $data['user_id'] = Auth::id();
@@ -34,16 +61,34 @@ class SettingService
         return Mail::to($item->tmp_email)->send(new VerifyNewEmailUserMail($item));
     }
 
-    public function updateEmailUser(ChangeEmail $tmp_item, User $item): User
+    /**
+     * @param string $key
+     * @return User
+     * @throws \Exception
+     */
+    public function updateEmailUser(string $key): User
     {
+        $tmp_item = $this->settingRepository->findTmpChangeEmail();
+        if (empty($tmp_item))
+            throw new \Exception(sprintf('An error occured. Start the procedure of changing the email address again'));
+        if ($tmp_item->_key != $key)
+            throw new \Exception(sprintf('Code provided is invalid'));
+        $item = $this->settingRepository->findUser();
+        if (empty($item))
+            throw new \Exception(sprintf('User not found'));
         $item->update(['email' => $tmp_item->tmp_email]);
         Mail::to($item->email)->send(new SuccessUpdateEmailMail());
         $tmp_item->delete();
         return $item;
     }
 
-    public function updateAvatar($avatar, $user_avatar)
+    /**
+     * @param $avatar
+     * @return mixed
+     */
+    public function updateAvatar($avatar)
     {
+        $user_avatar = $this->settingRepository->findAvatarUser();
         $b_path = Avatar::PATH;
         $file_name = md5(time() . Str::random(32)) . '.' . $avatar->getClientOriginalExtension();
         $avatar->move($b_path, $file_name);
